Update Israel
I apologize for the late submission.  I mistook the due date to be this upcoming Monday.  Would it be possible to have one place where information is updated.  My uci.edu email is hard to keep up with because of all of the campus list serves that I am on.  Let me know if you any questions.  Thank you